backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pytest
pytest-asyncio
python-jose[cryptography]
passlib
bcrypt==4.0.1
pydantic-settings
